chore: move some function into utils
